1. 解决量化模型配置eagle3投机时加载eagle模型卡住的问题 2. 解决解决量化模型配置eagle3投机时get_embedding_tensor检索出错的问题

+++ vllm_ascend/patch/worker/patch_draft_quarot.py
@@ -111,7 +111,7 @@ def make_load_weights(target_model_path, rotation_path):
             if "fc." in name:
                 # anti-rotate fc
                 dtype = loaded_weight.dtype
-                loaded_weight = loaded_weight.to("npu") @ Q3.to(dtype).to("npu")
+                loaded_weight = (loaded_weight.to(torch.float32).to("npu") @ Q3.to(torch.float32).to("npu")).to(dtype)
             if "embed_tokens" in name:
                 includes_embed_tokens = True
             model_weights[name] = loaded_weight
@@ -120,7 +120,9 @@ def make_load_weights(target_model_path, rotation_path):
         # process embedding if drafter does not have embedding
         if not includes_embed_tokens:
             name = "model.embed_tokens.weight"
-            loaded_weight = get_embedding_tensor(target_model_path).to(embed_dtype).to("npu") @ Q.T.to(embed_dtype).to("npu")
+            loaded_weight = (get_embedding_tensor(target_model_path).to(torch.float32).to("npu") @ Q.T.to(torch.float32).to("npu")).to(
+                embed_dtype
+            )
             model_weights[name] = loaded_weight
 
             includes_embed_tokens = True
